my_ai_docs_system/hub-app/requirements.txt
Authlib==1.4.0
fastapi==0.115.6
jose==1.0.0
passlib==1.7.4
pydantic==1.10.19
pytest==8.3.4
python-dotenv==1.0.1
python_jose==3.3.0
Requests==2.32.3
setuptools==75.6.0
SQLAlchemy==1.4.46
starlette==0.45.0
bcrypt==4.0.1
